User Seed Added, Mailer Previews Appear in views
Added a basic user seed just for the premise of appearing
as the intended user for the welcome email mailer.

Altered the welcome email html and text files from user.login to
user.name (as no password has been setup in the application for now)
to facilitate the generation of welcome emails.