Compact Workflows File to Test Effectiveness
Using https://github.com/ruby/setup-ruby to create
a far more simplified rubyonrails.yml file that only
runs a single job compared to the multiple previously ran.

Removed ruby version as with this setup it is not required
to specify the version with a preexisting .ruby-version file
that I have confirming the Ruby version to be 3.3.0.