Junctions Model and Migrations Created, Also Sidekiq Update
Created a Junction model with several fields attributed to it, migrated
into the schema, added an additional model field for difficulty via an
additional migration file.

Junction model is loosely real world road junction based as of now, however
flexibility remains as to the eventual decision on how this will look
in the application frontend.

Sidekiq initializer file added to config/initalizers to assist with
Sidekiq configuration with Redis, likely to expand this file further in the future.