Country Model and DB Relations for Country setup
Cities were previously setup to belong to a country,
this branch and in particular this commit and PR deals
with the establishment of the country model and its
incorporation into the database.

The Country model uses a string of name for the country name
and a string of code for the ISO 3166-1 country codes.