Links Added to Cities and Junctions, Views Readjusted
Links now exist from the homepage to the index pages for
cities and junctions.

Added junctions outside of being nested inside the cities
resources routes to allow the junctions path to run
succesfully.

Cut down on excess view code for the index pages of both
cities and junctions, displays the relevant seeds generated
in the views for each.