Junctions Seeds Added, Reconfigured Several Related Files
Added 6 real junctions to the seeds file that are populated
along with the single preexisting seed already in the junctions
index. Also reconfigured the junctions code block to make it suitable
the increased amount of junction seeds now populated.

Added junctions to its own individual resources in the routes and removed
it from the nested former status within the cities routes resources.

Junctions controller updated with added create action code to account for
when new junctions will be added in the future.

City model updated to allow junctions to be destroyed as part of the seed generate
command process.