Bootstrap and Related Gems Installed, First Bootstrap Frontend Added
Latest Bootstrap version gem, sassc-rails and cssbundling-rails
gems added and installed within the Gemfile.

Ran the installation command for Bootstrap to prepare and perform
the install for many new files and new directories.

This is not the complete installation process as there are several steps
also required to perform to complete Boostrap's installation.

Added a generic navbar available from Bootstrap's website to confirm it
is rendering in the views.