Finished Bootstrap Installation, Dropdown Navbar Working
Ran Rails Stimulus and Turbo installation
commands to ensure they are up and running in the application.
This updated the cable.yml file.

Added a application.scss file importing Bootstrap to ensure Bootstrap
styling is used going forward.

Added a line to the application layout file for a Bootstrap script, this
ensures that Boostrap is functional, and in such instances as the previously
added Boostrap navbar that its dropdown is now acting as intended, as a dropdown.