Junction Create Form Created, Updated Related Files
Created a Bootstrap Simple Form for creating a Junction,
along with a field for the location and a selector of
already existing cities for the city the junction and location
is attributed to.

Updated the junction model to include a junction attribute
for using junctions as a field in the form and in general
in the application going forward.

Not fully working as of this commit, as the junction
and city fields are viewable and editable but not
displaying in the frontend in the index view when
created.