Added Placeholder for Empty Junction Names in View, Added Links and Changed Text
Expression added to the ERB tag for junction names, in the case of created
junctions via the new junction form it is not rendering frontend the junction name.
Seeds are rendering the junction name frontend, therefore the expression including
the pipes to display a fallback message in the index page in the place of the junction
name covering for the lack of a rendered junction name.

Website navigation is far easier with links to the homepage and new junction form
available in the junction new and index pages respectively.

Text added to junctions index to make it more welcoming.