User Authentication First Part of Process Carried Out
Installed the latest version of the bcrypt gem for password security.

Completed a DB migration to add password fields to the DB
and the user model.

Updated the user controller to handle the user authentication implemented
and also further authentication code to be added.

Added four individual CRUD specified actions to a resources users setup
within the routes file.

Homepage updated with a link at the bottom to the users index page, can navigate
from the users index page to the new user link to signup.

Setup User new and index pages for handling new user
signups and the index showing all active users with their
name and ID.