Sessions Controller and View Setup, Several Controllers Updated
Created a Sessions controller to handle the login process and if a user
is logged in or not an to log out via the destroy method.

Updated routes to account for login and logout paths.

Simple form used for the login new session view page.

Updated confirmations and users controller to prevent authenticated
users from being able to access pages they are not supposed to gain
access to.

Added the login action to the confirmations controller for users
fully confirmed once they havecompleted the confirmation process
following sign up.