StrongMigrations, Chartkick and Groupdate Gems Added
Added strong_migrations gem due to many DB changes association, model
additions and model to model relationships upcoming, to better ensure
a useful check in the form of this gem is implemented to prevent dangerous
and erroring migrations.

Chartkick added due to my desire to integrate charts into this application,
currently only showing as "Loading..." where added into the homepage for showing
user grouped by when they were created.

Grouping mentioned above performed with the addition of the groupdate gem, that
as the gem name states, can group by times and longer formats, such as second, minute,
day and week.

Removed a build application.js file due to a name conflict with the same named
file in the javascript folder.