Many Changes Made, Hotwire/Turbo and Chart Working, Bootstrap Not Working
Made many hours of changes to the application.

Dependent upon changes in the application.html.erb file and whether the
javascript import tags line is in the file or not will affect whether Turbo
or Bootstrap will render in the views, both are not currently working at
the same time in any scenario I have tried.

Post scaffold created to run and test basic Turbo modal commands, currently
working however not as particular as a modal appearing in the centre of the
page and placing the emphasis on the modal (via fading the rest of the page).

Other changes have been made including new Turbo files and partials.