Unconfirmed Emails Accounted For, Users Edit, Update and Destroy also Implemented
Various methods for unconfirmed users have been added to the user model to
determine and differentiate if a user is trying to access with an email
that is signed up but not authenticated, therefore the non authenticated
email is stored and cannot be attempted repeatedly.

Updated the confirmations controller edit method to take into account if the user account
has been confirmed or not which determines the path that they will be taken on.

Authentication module updated with a authenticate user method for if they are
authenticated or not.

User controller added edit, destroy and update actions for users wishing to complete
any of these actions for their accounts. Also setup a Simple Form edit user account
page.

Also updated the routes to account for the three new user account actions.