Extending Relationship Between Junction and Posts models
Looking for posts to be posts created for junctions. Added
several lines to the post and junction models along with both
controllers updated in regards to the new relationship between
the models.

Also created a Junctions show page for displaying junctions
created, link created in the index page for each respective
junction entry.