Geocoding Gem Setup, Creating Records Successfully Locally
Installed the latest version (1.8.2) of the geocoder gem following added
it to the Gemfile. Gem also configured to use miles unit of length.

Geocoder gem setup due to relevance to the subject area, holds
value with it being able to locate the precise coordinates of
a specific location and/or city.

Added new Junctions migration file with new columns specific to geocoder,
migrated these int the DB and also updated the Junction model with inclusion
of Geocoder values.

Rails console allows Junction create queries to be ran.

Also added to the preexisting seed file conditional statement describing
what is put if the seed process is successful.

A new method conditional statement has been setup also in the seed file
for what should be put if the geocoding aspect of the seeding is unsuccesful,
along with a short else conditional display message if it is successful.