Redirects and Timing Attacks Accounted for, ActiveSession Setup
Authentication concern and sessions controller updated to store_location and
set paths related to the session respectively.

Timing attacks taken into consideration via implementation of a new Rails 7.1
method for the User class to determine how the authentication is decided via
attributes.

Sessions controller also updated again in this commit to find and authenticate
the user in the create method going beyond the previous just find user action.

Also setup the ActiveSession model and related files in order to understand
who the user is, this value can then also be stored in the database to keep
record of it. DB shcema migrated with references specific to the new
active_sessions setup. Furthermore, user model and authentication concern files
updated to account for active sessions.