Various Minor Application Changes
Updated posts model to put a limit on the amount of characters that can be
inputted for a junction, to prevent an infinite amount of characters for a junction
to be able to be made.

Started to assign models and eventually specific variables to users, have assigned
junctions to the user via the belongs_to method.

Swapped Homepage navbar Junctions index and new Junctions links with each other,
felt it was more appropriate to have the general almost hub like junctions index
page as one of the main standard appearing non dropdown navbar links as such.