Moved Header and New Footer into Partials, added CSS and JS Retuned
Header and footer partials created in the shared views folder.

Moved navbar code into a new header partial from the home view page,
rendered it into the applicationERB file so it and the navbar specifically
appears across the various views of the application.

New footer partial also created, this will deal with the likes of small
messages designed to be seen across the application and a few links to less
important pages than what exists in the header navbar.

Made adjustments to the application.js and manifest.js to ensure the
 rails assets:precompile command is working efficiently and without error.