First Passing RSpec Test
Created views folder and nested in views pages folder.
Within this folder a home RSpec test file has been setup
initially just to check that certain content renders in
the page.

Have tested with different strings and I can confirm that
with a string/text not on the homepage the RSpec test will fail.

RSpec test is currently passing.