Trying Phlex, Incorporated into a View Partial
Latest version of the phlex-rails gem installed. Completed
the install of phlex-rails which has created several files
and modified several other already existing files.

Generated a Phlex component file that is called in a new pages ERB partial modal.
The Phlex modal component (not currently performing as a modal however) is then
displayed accurately on the homepage via a rendering call of the modal partial
in the home.html.erb view page.

This does not affect the passing RSpec homepage test.