Updated User Controller and User Index View, Performed New DB Migration
Removed debugger from the user controller.

Updated the user index to render users names and links to the specific
users pages, which will then render the show page.

Performed new database migration for working towards resolving issues
with crossroads and cities so that the city_id is added to the crossroad
column in the schema in the form of an integer.